[op][n_upsable][cookies] Add per application cookies settings.
:Release Notes:
Allow to change the cookie security settings per application.

:Detailed Notes:
In last years Chromium has been disabling more and more cookies behavior
that could be considered dangerous, and has the goal of eventually
disabling third party cookies support in the future.

In webOS, for same reasons, we want to be able to only allow insecure
cookie usages when it is needed. For that purpose, this change allows
to set specific security settings for cookies per application. WAM will
also allow to tune the default behavior for the system.

The only settings offered is:
* SetAllowThirdPartyCookies (default: true). Allow third party cookies,
at all.

The default match the Chromium m91 defaults.

:Testing Performed:
Launching applications with different settings, and checking if the
cookies are allowed opening different test pages.

:QA Notes:
Depends on: http://gpro.lge.com/c/webosose/wam/+/333552
Test-cases: QWTC-618

:Issues Addressed:
[WRN-10227] WAM: third party cookies configuration support
[WRO-8372] Port [op][cookies] Cookies per app and default configuration.

Change-Id: I82042268a753c51933c8a9af257c4abf78872142
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/326435
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/333551
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>