[op][n_upsable][serviceworker] Integrated service worker openWindow to launch webapp
:Release Notes:
Integrated service worker openWindow to launch webapp.

:Detailed Notes:
A service worker can open window with a url and when it comes to webapp
we can consider to launch a webapp with given url.

To know the application id from service worker, we store application id
in service worker registration data so that it can be restored when
service worker is enabled and this information is used when openWindow
is called to launch the webapp.

:Testing Performed:
1. Prepare serviceworker onnotificationclick handler like
self.addEventListener('notificationclick', function(event) {
  event.notification.close();
  self.clients.openWindow("http://google.com/")
});

2. Launch webapp and close
3. Send push to webapp
4. When notification is showed up, click the notification.
Expected result: the app is launched with opening google.com page.

:QA Notes:
Test-cases: QWT-4520

:Issues Addressed:
[WRO-2149] Service worker for web push refactoring
[WRO-2052] Migrate the PoC result for service worker to chromium94
[WRN-11305] Consideration of application ID for web push

Change-Id: Iecf0e1b661e5a85a04ca611efcd89d946d7755b9
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/328774
Tested-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/328771
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>