[op][n_upsable][navigation] Fix not to invalidate surface to recover from crash
:Release Notes:
Fix not to invalidate surface to recover from crash

:Detailed Notes:
Feature of recovering from renderer process crash was broken after
a upstream patch:
==========
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2585790
commit 865a12d72b770520f88775d5e37810ead62d82c8
Author: Jonathan Ross <jonross@chromium.org>
Date:   Wed Dec 16 22:33:00 2020 +0000
Evict Surfaces from Before Navigation When Re-using DelegatedFrameHost
==========

After crash, the process was recovered but first_surface_activated_ was
false forever with the recovered process so the screen was never drawn.

And this was because of the above upstream patch that introduced a logic
that invalidates the surface while launching the app.
The first surface was invalidated so new surface id was created but
the new surface was not in the active_reference_surfaces_ so the logic
for finding the active surface was failed.

Removing the invalidation of the surface might not affect the fix of the
upstream patch since it actually evicts the surface with
pre_navigation_local_surface_id_ and
first_local_surface_id_after_navigation_.

:Testing Performed:
1. Launch an app
2. Kill the app (kill -9 [PROCESS_ID])
3. Verify that the app is recovered and is usable

:QA Notes:
Test-cases: QWT-417

:Issues Addressed:
[WRN-11767] After user guide is reloaded 'Getting started with
            LG webOS' page is not loading.

Change-Id: Ica6e9c4991c931618ca86d0a014f5a9806c5626a
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/326753
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/327687
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/332918
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>