[op][n_upsable][lttng] Avoid symbol clash with perfetto and LTTng.
:Release Notes:
Fix building webOS PmTrace support with recent Chromium implementation.

:Detailed Notes:
In Chromium m91, Perfetto libraries are used even if it the client
library is not used, because of the tracing library implementation.
But Perfetto library defines the some properties with the name
tracepoint, that clashes with the LTTng tracepoint define.

To avoid that problem, we removed dependency on lttng/tracepoint.h in
Chromium source code. With the change, we provided wrapper api for
tracing instead of exposing the LTTng tracepoint macro directly.

:Testing Performed:
1. Build enabling LTTng in WAM and Chromium.
2. Check that WebAppMgr exposes its tracepoints using lttng list -u

:QA Notes:
Depends on: http://gpro.lge.com/c/webos-pro/meta-lg-webos/+/326842
Test-cases: QWT-371

:Issues Addressed:
[WRN-12099] Strings are not displayed in Trace-file.
[WRN-14858] Port [op][n_upsable][lttng] Avoid synbol clash with perfetto
and LTTng.

Change-Id: I0535e33ec21c156a673c02b831c2b82585571a54
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/321834
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/326843
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/332882