[op][ds][splash] Add SetFirstActivateTimeoutMs method
:Release Notes:
Added SetFirstActivateTimeoutMs method for support splash dismiss
timeout.

:Detailed Notes:
This patch added SetFirstActivateTimeoutMs to webos::WebAppWindowBase to
support setting splash dismiss timeout.
This API provide a way to modify activate_eventually_timeout variable
which is used for timer to start BeginFrame of display if the surface
of renderer hasn't met the frame with FCP marked.
Previously it was constant value 8000ms and now it can be configurable
for the each app. Also the timer is started as soon as it was set so
that changing display visible starts BeginFrame immediately if the timer
has been expired.

:Testing Performed:
1. Add "splashDismissTimeoutMs": 1000 to appinfo.json
2. Start the app
3. Splash screen is displaying.
4. Splash screen should be removed within 1 second.

:QA Notes:
Test-cases: QWT-4477

:Issues Addressed:
[WRN-17952] Splash dismiss timer

Change-Id: I22f78cd408b9bde74eccf106590b29e4b99462b7
Follows-after: http://gpro.lge.com/c/webosose/chromium94/+/324787
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/327229
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Alexander Kovalenko <alexander.kovalenko@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Wanchang Ryu <wanchang.ryu@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/332887
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>