[op][ds][memory] Use tcmalloc
:Release Notes:
Avoid regression because of new partition allocator fragmentation by
using tcmalloc allocator.

:Detailed Notes:
The new partition allocator, that will provide many features related
to security and performance, is still problematic in m91 and m94.
On one side we are not getting all the advantages of it as raw
pointers are not using it. On the other side, fragmentation fixes are
still not available.

Because of all this, we are observing a serious regression on
repeatedly launching applications, compared to m87. In the future
tcmalloc will not be an option as code will be removed from upstream
tree. But, for m91 and m94 we should keep using tcmalloc.

:Testing Performed:
Launch media discovery 6000 times, and track VmRSS in procfs status
file for browser process.

:QA Notes:
Test-cases: gate review memory leak test.

:Issues Addressed:
[WRN-17221] [1st Gate Review] Memory leak

Change-Id: I40f4d85328a4c5a51086e5c2445362efba16cf27
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/329294
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Alexander Kovalenko <alexander.kovalenko@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/333001
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>