[op][n_upsable][static_analysis] Avoid time_t
:Release Notes:
Avoid time_t because the static analyzer says 32-bit time_t might cause
the year 2038 problem.

:Detailed Notes:
The problem will not happen because we just compute the time zone.

We use absl::Time to avoid the use of time_t.

:Testing Performed:
Compute webOSSystem.timeZone in a web application and check the result.

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRO-6563] [Static_Analysis][High_Impact] Use of 32-bit time_t

Change-Id: Ibb1a21d8fd30233c45af3ffcf5ce0a775ab1c38f
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/331525
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/332935
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>