[op][n_upsable][ozone][wayland] Terminate on error
:Release Notes:
Make WAM terminate when there is a Wayland error.

:Detailed Notes:
When ui::WaylandEventWatcher::CheckForErrors() gets a Wayland error
it tries to shutdown the web engine by running the shutdown callback.
Because the shutdown callback is currently not set,
WAM can be alive and do unexpected things after Wayland errors.
This patch sets the shutdown callback.

See ChromeBrowserMainExtraPartsOzone::PostCreateMainMessageLoop()
in //chrome/browser/chrome_browser_main_extra_parts_ozone.cc
for the upstream implementation.
Unlike the upstream, we set the exit code to 1 instead of 0,
because we want to make systemd restart WAM by indicating the failure.

:Testing Performed:
Stop fluent-bit.service to avoid making error reports.
Launch a web application.
Check the status of WAM before and after killing the LSM process.
The process id of WAM should be changed.

:QA Notes:
Test-cases: N/A

:Issues Addressed:
[WRO-6508] [RDX_CRASH] /usr/bin/WebAppMgr in n/a (libcbe.so + 0x3e7df30)
[WRO-7616] Terminate WAM on Wayland error

Change-Id: Ifa4b0fd07d1d5264edd075545f9d84dcd27eb7eb
Originally-Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/332524
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/333299
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Joonhyung Hwang <joonhyung.hwang@lge.com>
Tested-by: Wanchang Ryu <wanchang.ryu@lge.com>