Revert "Revert "[oe][n_upsable][wayland] Implement GPU vsync provider for wayland""
:Release Notes:
Recover Neva upstream patch that breaks build with Wayland
presentation. Follow-up patch provides fix.

:Detailed Notes:
As the reverted patch is coming already from Neva upstream, reland as
is to provide the follow up patch that can be upstreamed to Neva.

This reverts commit 1a646eda0e33f8241cd7cde5dd950e5ec5a85328.

:Testing Performed:
None. Patch to test is the head of the review chain.

:QA Notes:
Do not land alone, but with the follow up fix.

:Issues Addressed:
[WRO-8604] Reland [oe][n_upsable][wayland] Implement GPU vsync provider for wayland on GOW

Change-Id: I98a377b8ac983653fdaf243e877f8da665db06f6
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/333548
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>