[op][n_upsable][media] Fix suspend/resume with WebMediaPlayerImpl
:Release Notes:
Fix suspend/resume scenarios for background app with upstream
WebMediaPlayerImpl when using VDA

:Detailed Notes:
Currently while using VDA than media playback is not suspended when
the application goes to background. But with webOS framework we have
to suspend the media playback and unload the pipeline if the web app
goes to background and resume the playback when it become foreground
With this changes, we use upstream provided suspend/resume framework
to achieve the same.
Also suspend/resume implementation of NEVA media players has been
retained to function as it is by using runtime disable neva media
player flag.

:Testing Performed:
Launch an application with media content and play the video
After playback is started launch another application without media
Ensure that video playback is suspended and pipeline is unloaded
Re-launch the first application with media content.
Ensure that video playback is resumed at the same position.

:QA Notes:
QWTC-6833

:Issues Addressed:
[WRO-8991] : Youtube video is still playing after launching the
             bareapp
[WRO-12481] Patch migration (submissions/21 to submissions26) from ch91

Change-Id: Icc4d7a8d02deb0783b17a8d122cda40e7221b008
Reviewed-on: http://gpro.lge.com/c/webosose/chromium91/+/333721
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Karthick Sundaram <karthick.sundaram@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/335915
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>