[op][n_upsable][luna] Fix LunaServiceClient for utility process
:Release Notes:
Fix LunaServiceClient to receive luna callback in utility process

:Detailed Notes:
Currently luna response  callbacks are not received in utility
process because g_main_context_iteration is not running/executed
in this process.
So in this patch set we create a single thread task runner and do
PostTask to run a loop and and invoke g_main_context_iteration
till we receive the response callback for each luna call.
Since this is not needed for main process so we have added a new
arguement as run_gmain_context to LunaServiceClient c'tor with
default value as false.
The gmain_task_runner_ will be created and task will be posted
only when client pass true for the same.

Also fixed some of the repeated LOGs and rearranged the null check
for LSHandle.

:Testing Performed:
Build completed without any compilation error.
Functionality is tested along with and as mentioned in
https://gpro.lge.com/335626/

:QA Notes:
Test-cases: QWTC-5929, QWTC-5924

:Issues Addressed:
[WRO-13145] Fix responce in LunaServiceClient for utility process

Change-Id: Ie340e1085f4d8c8a3dfccbc1f191c39422dba979
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/337084
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Sujeet Nayak <sujeet.nayak@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>