[op][ups][i18n] Fix artifact on Sri Lankan character
:Release Notes:
Fix artifact on Sri Lankan character

:Detailed Notes:
This is a cherry-pick from upstream which fix the artifact found on
Sri Lankan character.

https://chromium-review.googlesource.com/c/chromium/src/+/3644742

Fix a bug with calculation of ShapeResultView's previous safe offset

A ShapeResultView's previous safe to break offset used to be
calculated by assuming that a part's first character corresponds to
the part's start index. For LTR text, however, the start index is
often, if not always, 0, meaning that for ShapeResultViews which start
at some non-zero index into the text (such as for lines in a paragraph
other than the first), the returned previous safe offset could be much
earlier than it should, or it could even be before the start of the
view.

Since NGLineBreaker::TruncateLineEndResult uses the previous safe
offset to determine which parts of a new line to reshape, if the
returned safe offset happens to be in the range of the view and in the
middle of a ligature, the first character in the ligature would not be
reshaped and would thus be set to the ligature glyph, while the
characters after the returned safe index would be reshaped and be set
to their corresponding glyphs. This leads to the appearence of the
characters after the offset being repeated.

This was fixed by changing `run_start` to be the part's start index
plus the view's char index offset.

Bug: 1304876
Change-Id: Ied68dd5986b603814fa236e7ab2081db5a817723
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3644742
Reviewed-by: Koji Ishii <kojii@chromium.org>
Commit-Queue: Andreu Botella <abotella@igalia.com>
Reviewed-by: Dominik Röttsches <drott@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1006370}

:Testing Performed:
1. Redirect to
   http://156.147.61.46:5002/webos22/tv_mr3/pc/v010301/index.html
2. Select "아주/중아" and "sin" and click OK button
3. Select the third item from the bottom
4. Select the third item (from the top)
5. Select the third item (from the top) again
6. Verify that there is no artifact on the character around "HDMI
   (eARC/ARC)". (See WRO-12378 for details)

:QA Notes:
Test-cases: QWTC-9822

:Issues Addressed:
[WRO-12378] [ch94] Certain Sri Lankan text has circle

Change-Id: Ia7ce064579459c7c75b6b6b7f95e457dc5189561
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/336330
Tested-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Youngsoo Choi <youngsoo.choi@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
(cherry picked from commit 293a0c5a14b8d1e3dc9faf1dbae17c5e178d4e51)
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/337526
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>