fixup! [op][ds][memory] Fix the problem with drawing of the window after sideloading on container-based apps
:Release Notes:
Fix expunging GPU memory of hidden applications.

:Detailed Notes:
After fixing the problems to restore side load of container based
applications, the memory expunging was not delaying compositor
suspend, and this caused a regression in memory.

When a delegated frame is expunged, if compositor is suspended, then
the memory is not freed. So it is required to suspend compositor
later.

:Testing Performed:
1. Launch media player.
2. Launch moonstone sampler.
3. Wait a few seconds.
4. Trace 5 seconds with memory-infra category enabled to get a
snapshot of cc memory allocation.
5. Check in memory trace capture, cc category of browser process.
There should be only one provider (the one for moonstone sampler).

:QA Notes:
Test-cases: QWTC-4013, QWTC-7213

:Issues Addressed:
[WRO-11010] The available memory of step 11 is bigger than the available memory of step 5.

Change-Id: Id3f495fd68567a22da3ed22bda98119e29c99ee6
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/336368
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Jose Dapena Paz <j.dapena@lgepartner.com>
(cherry picked from commit 58239914ee8542e2da5a0cae4669c7e795dda387)
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/337524
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>