fixup! [op][n_upsable][media] Implement WebOSVideoEncodeAccelerator for webOS
:Release Notes:
Fix for upstream changes related Bitrate and optimized scaling

:Detailed Notes:
In case of some resolution like 480x270 driver expects a diffrent
size input buffer like 512x270. In that case input buffers has to
be scaled to 512x270. Otherwise it fails either in Mojo Video
Encode Accelerator Service class Encode function or it does crash
when we push the buffers to the driver. So we use the input coded
size provided by the encoderfor doing optimized scaling.
In upstream bitrate arguement is changed to class type so we also
change the argument of RequestEncodingParametersChangeTask to same.
We have removed IsBitrateTooHigh since the same is already checked
in RTCVideoEncoder.

Also we have added FPS calculation code for tracking performance and
some error LOG adjustments.

:Testing Performed:
1. Start video call through MS Teams & join the call.
   Device A - Rpi4 with C270E
   Device B - Laptop
2. Local preview in Device A shoule be fine.
3. The Camera preview of Rpi4 should be fine in Device B(Laptop).

:QA Notes:
Test-cases: QWTC-7943

:Issues Addressed:
[WRO-15101] The green macro blocked data are streamed during video call

Change-Id: Ieadc2ee4efbdb0bf31ea0ddf4bcbfb26798dd002
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/338823
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>