[op][n_upsable][appshell] Add CORS exceptions to load pdf.js resources
:Release Notes:
Add CORS exception and register file scheme as web safe scheme to load
only PDF content.

:Detailed Notes:
File scheme is added as web safe scheme to allow loading of local
resources of pdf.js. The process id of the PDF web content is added as
CORS exception and the Web security preference is set as disabled for
the same pdf web content. So that when the web security is disabled
only the given exception list of the process ids are allowed for
CORS exception.

:Testing Performed:
1. Launch Enact browser and open any of these sample PDF urls.
https://www.africau.edu/images/default/sample.pdf
https://unec.edu.az/application/uploads/2014/12/pdf-sample.pdf
https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf
2. You should be able to scroll through the PDF page and do Zoom in/out
3. Open Google search and search for 'sample PDF download' and click on
any pdf page links, the PDF page should get loaded.

:QA Notes:
Test-cases: QWTC-11363

:Issues Addressed:
[WRP-7605] Implementation on PDF viewer and creating bundles of the
PDF.js prebuilt library at the Enact side

Change-Id: Ic68716fb927ddcc57602416485ea1fea55792090
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/340741
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>