[op][n_upsable][app_shell] Make windows translucent by default
:Release Notes:
Made windows spawned on-behalf of the app_shell binary being
of the translucent opacity by default.

:Detailed Notes:
Since Qt 6.4.0 alpha-blending for an opaque (non-transparent)
surface is forcibly disengaged due to
https://codereview.qt-project.org/c/qt/qtwayland/+/395300.

With the above change, when trying to play a video from the
HTML5 tag rendered on-behalf of the app_shell binary (e.g.,
the Enact-based browser), the video becomes invisible when
played.

The issue is caused by the wl_surface_set_opaque_region()
request sent due to the root surface was set as fully
opaque (a default setting for the window trait when created).

On the other hand, windows spawned on-behalf of WAM are
always translucent by default (under the hood of the Neva
app-runtime). So, the issue is not reproduced when trying
to play the video from a WAM-managed web-app.

Thus, the changelist aligns assumptions on traits for the
spawned windows between both WAM and app_shell perspectives
and fixes the issue.

:Testing Performed:
Steps to reproduce:
1. Launch the Enact-based browser from the console:
   $ sh -x /usr/bin/app-shell/run_app_shell
     --folder /usr/palm/applications/com.webos.app.enactbrowser/
     --appid com.webos.app.enactbrowser

2. Open the following URL Within the browser's omnibox:
   https://www.w3.org/2010/05/video/mediaevents.html

3. Start playing a movie from the embedded video player

4. Ensure that the played movie is visible
   (no blackscreen instead of a played movie is observed)

:QA Notes:
Test-cases: QWTC-492

:Issues Addressed:
[WRP-3204] [OSE-Emulator] : HTML5 video not visible
           in browser.

Change-Id: I887e64c9de93927f7004eeabea806fc1c97837e0
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/341988
Tested-by: Sergey Kipet <sergey.kipet@lge.com>
Tested-by: Test BTS <test.bts@lge.com>
Reviewed-by: Oleg Chaklyarov <oleg.chaklyarov@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Vladislav Mukulov <vladislav.mukulov@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>