[oe][n_upsable][audio] Filter audio device enumeration for display
:Release Notes:
Filter Audio devices list based on display for device enumeration

:Detailed Notes:
Filter list of audio devices associated with the display id on
which the application is running that requested the list of devices
In OSE audio service provides the display id information for which
a device is meant to be used. Also from the Renderer prefs we know
on which display the application is running. In MediaDevicesManager
we match these information to filter the list of devices.

This is needed to avoid unexpected behavior when user selects a
device (meant for display-2) while the application is running on
display-1

Also we need to pass the actual device id to the pulse audio when
user selects a device from the list along with the stream type.
That is why the extra arguement is added in pulse_util functions.

:Testing Performed:
Build is successfull for all platforms.
1. Install test.redirection.webrtc which allows audio/video capture
2. Lunch with below targte url:
   https://webrtc.github.io/samples/src/content/devices/input-output/
3. Camera preview should be visible on the screen
4. Change the device from Audio input source list
   - Ensure that there is no issues and mic is accessible
5. Change the device from Audio output destination list
   - Ensure that sound from mic is audible without any issues

:QA Notes:
Test-cases: QWTC-5925, QWTC-5924

:Issues Addressed:
[WRP-3366] Filter audio device enumeration specific to display id

Change-Id: I518a464802e952ed2fe3af906f80711c4896a65e
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/342231
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>