[oe][n_upsable][lttng][clang] Restore work of lttng provider
:Release Notes:
Tracing with the custom lttng provider was restored.

:Detailed Notes:
Custom chromium_lttng_provider didn`t work if it was compiled and linked
with clang into separate dynamic loaded liblttng_provider.so. It`s
happens because the init ptrs
lttng_ust_tracepoint_ptr_##_provider##___##_name of the
chromium_lttng_provider is optimized out during linking stage regardless
of the presence of the __attribute__((used)). It is happens because of
the clang specific behaviour - needed the additional "retain" attribute
for the ELF target when the "-ffunction-sections -fdata-sections" and
"-Wl,--gc-sections" is present simultaneously. See
https://clang.llvm.org/docs/AttributeReference.html#used
Optimization of the loadable module was disabled to prevent garbage
collection.

:Testing Performed:
Test 1 - build test
1. build libcbe.so with the use_lttng=true (
$ bitbake webruntime-clang
2. Check for the symbols of the lttng provider with the {cross-toolchain}-nm
the __start* and __stop* symbols should be defined:
$ cd {CHROMIUM_SRC}
$ ./src/third_party/llvm-build/Release+Asserts/bin/llvm-nm -C ./build/Release/libcbe.so | grep lttng_ust_tracepoints_ptrs
000000000ab40960 d __start_lttng_ust_tracepoints_ptrs
000000000ab409b0 d __stop_lttng_ust_tracepoints_ptrs

Test 2 - runtime test
1. Connect to the target board via ssh
2. Create and start new LTTng session
$ lttng create && lttng enable-event -u -a && lttng start
3. Restart WAM
$ restart webapp-mgr
4. Stop LTTng session
$ lttng stop
5. Check traces. chromium_lttng_provider should be traceable
$ lttng view | grep cromium_lttng_provider
 [03:31:23.008860884] (+0.000019704) raspberrypi4-64 chromium_lttng_provider:scope_exit: { cpu_id = 1 }, { category = "base", scope = "WorkerThread active" }

:QA Notes:

:Issues Addressed:
[WRP-9320] Verify and fix the work of lttng

Change-Id: I02f0084558311f0108c00fab0fe87adc43a44711
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/353168
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>