fixup! [oe][media][build] Preventing unresolved media calls
:Release Notes:
Preventing unresolved media calls

:Detailed Notes:
This problem is hidden in case of the lld usage and use of the
--gc-sections linker option.
But removing the optimization link option lead to correct, non-optimal
linker behaviour and it try to build binary with some unused classes
and methods. They are belong to the neva_media and have a not obvious
and inaccurate dependecies with the upstream chromeos media src.
This issue is hidden with the --gc-sections optimization option enabled,
but is not resolved. There are no symbols of the caller/called classes
(which cause this issue) in the webruntime lib.
Alternate way is to solve all inaccurate dependecies from media
component. But upstream changes already present in the NEVA-7748 (use
WebOS VDA/VDEA instead of the neva_v4l2_codec). Current patch should be
revised (refactored or removed) after merge those NEVA changes.

:Testing Performed:
Build only

:QA Notes:
N/A

:Issues Addressed:
[WRP-9320] Verify and fix the work of lttng

Change-Id: I38bbd4450a0238dc2968ce1188a149acefcc2ae0
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/354310
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jose Dapena Paz <j.dapena@lgepartner.com>
Reviewed-by: Alexander Kovalenko <alexander.kovalenko@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Donghyun Kim <donghyun11.kim@lge.com>