fixup! [op][n_upsable][appshell] Fix issue for screen layout error when using VKB
:Release Notes:
Fix issue for screen layout error when using VKB

:Detailed Notes:
When VKB showing, shiftContent event is fired with the calculated height
of VKB then page content will be shifted.
If the user starts input by HID, the showing VKB will be hidden and
content must be shifted back, an event also dispatched with the height
is 0.
The problem happens if the user input by HID keyboard before the VKB
showing's animation finished (~600ms according the timer delay time)
because this makes the wrong order of the above events (shift back
run first).
To solve this, we must stop all shiftContent events if we receive
the shift back event.

:Testing Performed:
1. Launch Webex from the Enact browser APP and join the meeting
2. Touch the Chat input box(VKB output)
3. Keyboard Typing (use HID Keyboard device) when VKB is showing
4. Click chat input box (use HID mouse device)
5. Screen layout should be normal.

:QA Notes:
Test-cases: QWTC-12793

:Issues Addressed:
[WRP-15783] Screen layout error when repeatedly selecting chat input box/keyboard input

Change-Id: I12b7d5ec1dd1afb81a6cbc615a172b7584edc9f3
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/355626
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: <luc2.tran@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: <chuong3.nguyen@lge.com>