fixup! [op][n_upsable][appshell] Fix the issue with the tooltip not displaying
:Release Notes:
Fix the issue with the tooltip not displaying

:Detailed Notes:
The problem is that the tooltip is not being shown.
This feature was not added in the app-shell side.
To solve this, we create a tooltip object
and assign it to the root window to listen for mouse move events.
Note that the ViewsDelegate interface is not implemented in the appshell
side, which leads to a crash issue when initializing tooltips.

:Testing Performed:
1. Run Enact browser Application
2. Open youtube link. https://youtube.com
3. Move mouse focus to "All" button
4. Tooltip will show under cursor

:QA Notes:
Test-cases: QWTC-12840

:Issues Addressed:
[WRP-15770] Description of selectable tabs No speech bubble

Change-Id: I75d7d671e0314d2b4fab4cb16b70480c99366c83
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/357816
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Van Luc Tran <luc2.tran@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Van Luc Tran <luc2.tran@lge.com>