fixup! [op][n_upsable][webrtc] Implement web media player for webrtc
:Release Notes:
Fix crash, video visibility and SW fallback issue when player object
is not available

:Detailed Notes:
In webrtc pass through decoder, we use the first frame to ensure
video window creation and initialization of video window and media
platform api.
But sometimes in webex and ms teams, web media player is created
after first transparent frame is returned to decode complete callback
for player init.
In this case player does not receive the call and the frame is
returned from MediaStreamVideoTrack::FrameDeliverer::DeliverFrameOnIO
since callbacks_ has not been added by the player yet.
So, to overcome this situation, we stored the video frame there and
returned the same when the callback from web media player is added.

Again, in case of MS teams particularly when local camera is stopped
than first time once the current pass through decoder is destroyed and
new decoder instance is used. So, we rearanged the call sequence in
WebMediaPlayerWebRTC::HandleVideoFrame to pass the callback function
pointers and window id to new decoder instance to create andinitialize
the media platform api for new decoder.

Also added condition to check if player object is null in web media
player ms for crash issue fix.

:Testing Performed:
1. Launch WebRtc test app (com.webos.app.test.webrtc) with URL-
   https://webrtc.github.io/samples/src/content/peerconnection/change-codecs/
2. Start the local preview and click Call button after selecting
   H264 codec from the list.
3. Check that the camera content is displayed in both (left and
   right) video windows properly

:QA Notes:
Test-cases: QWTC-509, QWTC-1377

:Issues Addressed:
[WRP-23308] [OSE -RPI4]: Device B video is not visible in Device A display
[WRP-21754] [RDX_CRASH][webos] /usr/bin/app-shell/app_shell in n/a (libcbe.so + 0x5fb2b74)

Change-Id: I19f25aa90970feabe2b76beea1a7d04ea60f340a
Reviewed-on: http://gpro.lge.com/c/webosose/chromium94/+/364129
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Wanchang Ryu <wanchang.ryu@lge.com>
Reviewed-by: Pankaj Maharana <pankaj.maharana@lge.com>
Reviewed-by: Donghyun Kim <donghyun11.kim@lge.com>
Tested-by: Pankaj Maharana <pankaj.maharana@lge.com>